% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rminiconda.R
\name{find_miniconda_pip}
\alias{find_miniconda_pip}
\title{Find the pip binary executable for an rminiconda installation}
\usage{
find_miniconda_pip(name = "general", path = get_miniconda_path())
}
\arguments{
\item{name}{The name of the miniconda installation.}

\item{path}{The base directory where all "rminiconda" miniconda installations are located.}
}
\description{
Find the pip binary executable for an rminiconda installation
}
