% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_miniconda_python}
\alias{find_miniconda_python}
\title{Find the python binary executable for an rminiconda installation}
\usage{
find_miniconda_python(name = "general", path = get_miniconda_path())
}
\arguments{
\item{name}{The name of the miniconda installation.}

\item{path}{The base directory where all "rminiconda" miniconda installations are located.}
}
\description{
Find the python binary executable for an rminiconda installation
}
