% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rminiconda.R
\name{get_python_version}
\alias{get_python_version}
\title{Get the Python version of a miniconda installation}
\usage{
get_python_version(name, path = get_miniconda_path())
}
\arguments{
\item{name}{The name of the miniconda installation.}

\item{path}{The base directory where all "rminiconda" miniconda installations are located.}
}
\description{
Get the Python version of a miniconda installation
}
