% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_miniconda}
\alias{install_miniconda}
\title{Install miniconda}
\usage{
install_miniconda(version = 3)
}
\arguments{
\item{version}{The major version number of Python (2 or 3). The latest version of the specified major version will be installed.}
}
\value{
\code{NULL} (miniconda is installed to a system directory).
}
\description{
Download the miniconda installer and run it.
}
