% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rminiconda.R
\name{is_miniconda_installed}
\alias{is_miniconda_installed}
\title{Check if an rminiconda installation exists}
\usage{
is_miniconda_installed(name = "general", path = get_miniconda_path())
}
\arguments{
\item{name}{The name of the miniconda installation.}

\item{path}{The base directory where all "rminiconda" miniconda installations are located.}
}
\description{
Check if an rminiconda installation exists
}
