% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_miniconda}
\alias{remove_miniconda}
\title{Remove an "rminiconda" miniconda installation}
\usage{
remove_miniconda(name = "general", path = get_miniconda_path())
}
\arguments{
\item{name}{The name of the miniconda installation.}

\item{path}{The base directory where all "rminiconda" miniconda installations are located.}
}
\description{
Remove an "rminiconda" miniconda installation
}
