% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rminiconda.R
\name{rminiconda_pip_install}
\alias{rminiconda_pip_install}
\title{Utility function to pip install a package in an rminiconda installation}
\usage{
rminiconda_pip_install(pkg_name, name, args = "")
}
\arguments{
\item{pkg_name}{The name of the pip package to install.}

\item{name}{The name of the miniconda installation.}

\item{args}{Optional string specifying additional arguments to pip, e.g. "-i https://test.pypi.org/simple/"}
}
\description{
Utility function to pip install a package in an rminiconda installation
}
