% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rminiconda.R
\name{rminiconda_pip_uninstall}
\alias{rminiconda_pip_uninstall}
\title{Utility function to pip unininstall a package in an rminiconda installation}
\usage{
rminiconda_pip_uninstall(pkg_name, name, args = "")
}
\arguments{
\item{pkg_name}{The name of the pip package to unininstall.}

\item{name}{The name of the miniconda installation.}

\item{args}{Optional string specifying additional arguments to pip}
}
\description{
Utility function to pip unininstall a package in an rminiconda installation
}
