% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rminiconda.R
\name{test_miniconda}
\alias{test_miniconda}
\title{Run a simple "hello world" test of a miniconda installation}
\usage{
test_miniconda(name, path = get_miniconda_path())
}
\arguments{
\item{name}{The name of the miniconda installation.}

\item{path}{The base directory where all "rminiconda" miniconda installations are located.}
}
\description{
Run a simple "hello world" test of a miniconda installation
}
